fix genesis and related proto

+++ x/register/keeper/params.go
@@ -26,7 +26,7 @@ func (k Keeper) BondDenom(ctx sdk.Context) (res string) {
 
 // MaxEntries - Maximum number of simultaneous unbonding
 // delegations or redelegations (per pair/trio)
-func (k Keeper) MaxEntries(ctx sdk.Context) (res uint16) {
+func (k Keeper) MaxEntries(ctx sdk.Context) (res uint32) {
 	k.paramSpace.Get(ctx, types.KeyMaxEntries, &res)
 	return
 }
